pyproject.toml
[project]
name = "flask-publicly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=3.1.3",
    "flask-login>=0.6.3",
    "gunicorn>=25.3.0",
    "mysql-connector-python>=9.6.0",
]
